+++ nodesafari/__init__.py
@@ -56,4 +56,4 @@ __all__ = [
     "top_link_predictions",
 ]
 
-__version__ = "1.2.0"
+__version__ = "1.2.1"